Refactor audio system and add notation toggle
