+++ index.html
@@ -1161,6 +1161,9 @@
               <button class="action-button ghost" type="button" id="color-scheme-button">
                 Palette · Aurora
               </button>
+              <button class="action-button ghost" type="button" id="notation-toggle-button">
+                Notation · Letters
+              </button>
             </article>
             <article class="card">
               <h3>Field Notes</h3>
